+++ Models/Models/Offer.cs
@@ -8,7 +8,7 @@ namespace Models
 {
     public class Offer : Post
     {
-        public Offer(int amount, double price, Currency currency) : base(amount, price, currency)
+        public Offer(double amount, double price, Currency currency) : base(amount, price, currency)
         {
 
         }
